Began Implementing Melee
Has box trigger in front of player that rotates towards mouse. Any enemies in this trigger are added to a list. When attack button (space) is hit, every enemy in the list will run their Hit() function (not written yet).

So far the cube and adding enemies to the list has been implemented. The cube however needs tweaking to not rotate on X and Z axis.